FarofaTrip/requirements.txt
asgiref==3.10.0
coverage==7.11.3
Django==5.2.7
django-cors-headers==4.9.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
